Extracting Structs
Extracted most structs to their own files.
