+++ Assets/Resorces/Scripts/Structs.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 736651cf9e9343642a03acc83133680f
+folderAsset: yes
+timeCreated: 1458851011
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
