+++ Assets/Resorces/Scripts/Structs/Animations.cs
