+++ Assets/Resorces/Scripts/Structs/Animations.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c476fc5b172d9c6479b0edebb4c467b8
+timeCreated: 1458851011
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
