+++ Assets/Resorces/Scripts/Structs/Direction.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+public enum Direction : int{
+  UP = 0,
+  DOWN = 2,
+  LEFT = 3,
+  RIGHT = 1
+}
