+++ Assets/Resorces/Scripts/Structs/GridLocation.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: fde81b6963ac73a4b82652d418e838f9
+timeCreated: 1458851014
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
