+++ Assets/Resorces/Scripts/Structs/OptionsInfo.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+using System;
+using System.Xml.Serialization;
+
+[XmlRoot("options")]
+[Serializable]
+public struct OptionsInfo{
+  [XmlAttribute("resolution")]
+  public Resolution resolution;
+}
