+++ Assets/Resorces/Scripts/Structs/OptionsInfo.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 867527d80fccb2247b9d12ac5de7a82c
+timeCreated: 1458851901
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
