+++ Assets/Resorces/Scripts/Structs/SerializedCompletedAction.cs
