+++ Assets/Resorces/Scripts/Structs/SerializedCompletedAction.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 7e9aa5728b24dac40a540a68d942ece5
-timeCreated: 1458843737
+guid: b44ce1a54202da347a792b599eff4538
+timeCreated: 1458851011
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
