+++ Assets/Resorces/Scripts/Structs/UnitSaving.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 82aebe8ce94c6144592d20a0b4960c52
+timeCreated: 1458851901
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
