+++ Assets/Resorces/Scripts/Structs/UnitTimer.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+[System.Serializable]
+public struct UnitTimer{
+  public float time;
+  public float maxTime;
+  public float ticAmount;
+}
