+++ Assets/Resorces/Scripts/Structs/UnitTimer.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 9e3082d0fc1a4254f92f62580fc3b093
+timeCreated: 1458851901
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
